Update WFAS mosaics only in presence of new files.

+++ bin/update_wfas.sh
@@ -15,13 +15,14 @@ EXT=('_perc_new' '_perc_new' '_new' '')
 
 function remove_files_from_mosaic {
 	#Get a list of coverages for this mosaic:
-	COVERAGES=(`curl -s -u admin:geoserver -XGET ${REST_URL}/${WORKSPACE}/coveragestores/${1}/coverages.xml \
+	IFS=$'\n' COVERAGES=(`curl -s -u admin:geoserver -XGET ${REST_URL}/${WORKSPACE}/coveragestores/${1}/coverages.xml \
 		                     |grep -oP '(?<=<name>).*?(?=</name>)'`)
 	for c in ${COVERAGES[@]}
 	do
+	   encoded=$(python -c "from urllib.parse import quote; print(quote('''$c'''))")
 	   #delete all granules:
 	   curl -s -u admin:geoserver -XDELETE \
-		"${REST_URL}/${WORKSPACE}/coveragestores/${1}/coverages/${c}/index/granules.xml"
+		"${REST_URL}/${WORKSPACE}/coveragestores/${1}/coverages/${encoded}/index/granules.xml"
 	done
 }
 counter=0
@@ -32,25 +33,34 @@ do
       mkdir -p $WFAS_DIR/${v}/tif
    fi
 
-   remove_files_from_mosaic ${v}
-   #remove geotiffs from previous forecast:
-   rm ${WFAS_DIR}/${v}/tif/*.tif*
    
    for d in `seq 0 6`
    do
       if curl -s -I ${REMOTE_ADDR}/${v}_day${d}${EXT[${counter}]}.tif; then #check that remote file exists
-         date=`date +'%Y%m%d' -d '+'${d}' days'`
          wget -q -N ${REMOTE_ADDR}/${v}_day${d}${EXT[${counter}]}.tif \
-            -O ${WFAS_DIR}/${v}/tif/${v}_${date}.tif ;
+            -P ${WFAS_DIR}/${v}/ ;
+	 modified=`stat ${WFAS_DIR}/${v}/${v}_day${d}${EXT[${counter}]}.tif |grep Modify|cut -d ' ' -f 2`
+         date=`date +'%Y%m%d' -d "${modified} +"${d}" days"`
+	 mv ${WFAS_DIR}/${v}/${v}_day${d}${EXT[${counter}]}.tif ${WFAS_DIR}/${v}/${v}_${date}.tif
+
+
       fi
    done
-   #now reindex the mosaic:  
-   for file in `ls ${WFAS_DIR}/${v}/tif/*.tif`
-   do
-	curl -s -u admin:geoserver -XPOST -H "Content-type: text/plain" \
-		-d "file://${file}" \
-		"${REST_URL}/${WORKSPACE}/coveragestores/${v}/external.imagemosaic" ;       
-   done
+   new_files=`ls ${WFAS_DIR}/${v}/${v}_*.tif|wc -l`
+   if [ ${new_files} = 7 ]; then
+	echo "New files: ${new_files}"
+   	remove_files_from_mosaic ${v}
+   	#remove geotiffs from previous forecast:
+   	rm ${WFAS_DIR}/${v}/tif/*.tif*
+	mv ${WFAS_DIR}/${v}/${v}_*.tif ${WFAS_DIR}/${v}/tif/.
+  	#now reindex the mosaic:  
+   	for file in `ls ${WFAS_DIR}/${v}/tif/*.tif`
+   		do
+		curl -s -u admin:geoserver -XPOST -H "Content-type: text/plain" \
+			-d "file://${file}" \
+			"${REST_URL}/${WORKSPACE}/coveragestores/${v}/external.imagemosaic" ;       
+  	 done
+   fi
    (( counter++ ))
 done
